+++ Pelituotanto/src/wildwestshootout/entity/Entity.java
@@ -27,7 +27,7 @@ import wildwestshootout.level.Level;
  */
 public class Entity {
     
-    protected int x, y;
+    protected double x, y;
     protected Sprite sprite;
     private boolean removed = false;
     protected Level level;
@@ -47,7 +47,7 @@ public class Entity {
     
     public void render(Screen screen) {
         if (sprite != null) {
-            screen.renderSprite(x, y, sprite, true);
+            screen.renderSprite((int) x,(int) y, sprite, true);
         }
     }
     
@@ -56,11 +56,11 @@ public class Entity {
         removed = true;
     }
     
-    public int getX() {
+    public double getX() {
         return x;
     }
     
-    public int getY() {
+    public double getY() {
         return y;
     }
