+++ Pelituotanto/src/wildwestshootout/entity/spawner/Spawner.java
@@ -32,7 +32,7 @@ public abstract class Spawner extends Entity {
     
     private Type type;
     
-    public Spawner(int x, int y, Type type, int amount, Level level) {
+    public Spawner(double x, double y, Type type, int amount, Level level) {
         init(level);
         this.x = x;
         this.y = y;
